requirements.txt
flask==2.3.3
discord.py==2.3.2
gunicorn==21.2.0
